chore: refactor: move defaults.ts into /lib, since it is referenced from client-side.
