+++ apps/app/src/electron/SuperConductor.ts
@@ -22,7 +22,7 @@ import { LoggerLike } from '@shared/api'
 import { hash, listAvailableDeviceIDs, rateLimitIgnore, updateGroupPlayingParts } from '../lib/util'
 import { findAutoFillResources } from '../lib/autoFill'
 import { Part } from '../models/rundown/Part'
-import { getDefaultPart } from './defaults'
+import { getDefaultPart } from '../lib/defaults'
 import { TimelineObj } from '../models/rundown/TimelineObj'
 import { postProcessPart } from './rundown'
 import { assertNever } from '@shared/lib'
