+++ apps/app/src/electron/storageHandler.ts
@@ -6,7 +6,7 @@ import { omit } from '@shared/lib'
 import { Project } from '../models/project/Project'
 import { Rundown } from '../models/rundown/Rundown'
 import { AppData } from '../models/App/AppData'
-import { getDefaultAppData, getDefaultGroup, getDefaultProject, getDefaultRundown } from './defaults'
+import { getDefaultAppData, getDefaultGroup, getDefaultProject, getDefaultRundown } from '../lib/defaults'
 import { ResourceAny } from '@shared/models'
 import { baseFolder } from '../lib/baseFolder'
 import * as _ from 'lodash'
